refactor(auth): replace TokenManager/LogoutManager with logout.utils
