+++ packages/shared/src/features/auth/index.ts
@@ -1,21 +1,11 @@
 // Types
-export type {
-  ExchangeResponse,
-  RefreshTokenResponse,
-  TokenPair,
-  AuthProfile,
-  AuthState,
-  TokenUpdatePayload,
-} from './types';
+export type { AuthSessionResponse, RefreshTokenResponse, AuthState } from './types';
 
 // Store
 export { useAuthStore } from './stores';
 
 // API
 export { exchangeTokenAPI, refreshTokenAPI, logoutAPI } from './api';
-
-// Hooks
-export { useTokenExpiration } from './hooks';
 
 // Utils
 export { LogoutManager, createLogoutManager } from './utils';
